TIP-860: Move repositories from enrich bundle to the right BC
