+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/repositories.yml
@@ -138,3 +138,17 @@ services:
             - '@pim_catalog.object_manager.product'
             - '%pim_catalog.entity.product.class%'
             - '@pim_catalog.repository.product'
+
+    pim_enrich.repository.category:
+        class: 'Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Repository\InternalApi\CategoryRepository'
+        arguments:
+            - '@pim_user.context.user'
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.category.class%'
+
+    pim_enrich.repository.group:
+        class: 'Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Repository\InternalApi\GroupRepository'
+        arguments:
+            - '@pim_user.context.user'
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.group.class%'
